# Proyecto 1/Backend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: SmartEgg API
  version: "2.0.0"
  description: |
    API REST de la plataforma IoT SmartEgg (Proyecto 1).

    Este documento define la interfaz que consume el frontend. Es la referencia
    unica de rutas, cuerpos, respuestas y codigos de estado.

    ## Separacion REST / MQTT

    | Va por REST (este documento) | Va por MQTT (broker) |
    |---|---|
    | CRUD de lotes, operadores y usuarios | Telemetria continua del nodo |
    | Consultas historicas y estadisticas | Alarmas criticas en tiempo real |
    | Login, roles y perfiles | Estado de actuadores |
    | Generacion de reportes PDF | Disponibilidad del nodo (LWT) |

    Las alarmas **no** se consultan por polling HTTP: el frontend se suscribe a
    `smartegg/{node_id}/alarmas/critica` por WebSocket (puerto 9001).
    El catalogo de topics vive en `GET /api/diagnostico/mqtt/topics`.

    ## Autenticacion

    1. `POST /api/auth/login` devuelve un JWT.
    2. Enviarlo en cada peticion protegida: `Authorization: Bearer <token>`.

    Roles: `administrador`, `operador_1`, `operador_2`.
    Un operador solo puede consultar **sus propios** diagnosticos e informes;
    el administrador puede consultar los de cualquiera.

    Los endpoints de telemetria de solo lectura son publicos a proposito, para
    que Grafana y las visualizaciones (p5.js / Chart.js) consuman datos reales.
servers:
  - url: http://localhost:5000
...
  - url: http://backend:5000
...
security:
  - BearerAuth: []
paths:
...
  /api/openapi.json:
    get:
      tags: [Sistema]
      summary: Esta misma especificacion, en JSON
      description: Lo consume Swagger UI. Publico.
      security: []
      responses:
        "200":
          description: Documento OpenAPI 3.0
          content:
            application/json:
              schema: { type: object }
        "500":
          description: La especificacion no se pudo leer del disco
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
...
  /api/auth/login:
    post:
      tags: [Autenticacion]
      summary: Iniciar sesion y obtener el token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: admin }
                password: { type: string, format: password, example: admin123 }
      responses:
        "200":
          description: Credenciales correctas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RespuestaLogin" }
        "400": { $ref: "#/components/responses/ErrorValidacion" }
        "401":
          description: Usuario o contrasena incorrectos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "403":
          description: La cuenta esta desactivada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "503": { $ref: "#/components/responses/BaseNoDisponible" }
...
  /api/perfil:
    get:
      tags: [Perfil]
      summary: Perfil del usuario autenticado
      description: Usuario del token junto con su ficha de operador y sus permisos.
      responses:
        "200":
          description: Perfil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Perfil" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
...
  /api/sensores/actual:
    get:
      tags: [Sensores]
      summary: Ultima lectura del nodo
      description: |
        Publico. Toma la lectura del hilo serial; si aun no hay ninguna en
        memoria, cae a la ultima guardada en la base.
      security: []
      responses:
        "200":
          description: Lectura mas reciente
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lectura" }
        "404":
          description: Todavia no se ha recibido ninguna lectura
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
...
  /api/lotes:
    get:
      tags: [Lotes]
      summary: Listar lotes
      description: Publico, para que el frontend pinte las tablas.
      security: []
      responses:
        "200":
          description: Lotes registrados
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lote" }
    post:
      tags: [Lotes]
      summary: Registrar un lote
      description: Requiere token.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoteEntrada" }
      responses:
        "201":
          description: Lote creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lote" }
        "400": { $ref: "#/components/responses/ErrorValidacion" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "503": { $ref: "#/components/responses/BaseNoDisponible" }
  /api/lotes/{lote_id}:
    parameters:
      - name: lote_id
        in: path
        required: true
        description: ObjectId del lote.
        schema: { type: string, example: 66f0a1b2c3d4e5f601234567 }
    get:
      tags: [Lotes]
      summary: Obtener un lote
      security: []
      responses:
        "200":
          description: Lote
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lote" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
    put:
      tags: [Lotes]
      summary: Actualizar un lote
      description: |
        Requiere token. Campos aceptados: `nombre_lote`, `cantidad_huevos`,
        `raza`, `responsable`, `observaciones`, `estado`. El `codigo` no se cambia.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoteActualizacion" }
      responses:
        "200":
          description: Lote actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lote" }
        "400": { $ref: "#/components/responses/ErrorValidacion" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
    delete:
      tags: [Lotes]
      summary: Eliminar un lote
      description: Requiere token.
      responses:
        "200":
          description: Lote eliminado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "404": { $ref: "#/components/responses/NoEncontrado" }

  # ----------------------------------------------------------- Operadores ---
  /api/operadores:
    get:
      tags: [Operadores]
      summary: Listar operadores
      description: Requiere token (cualquier rol).
      responses:
        "200":
          description: Operadores registrados
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Operador" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
    post:
      tags: [Operadores]
      summary: Registrar un operador
      description: Solo administrador.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OperadorEntrada" }
      responses:
        "201":
          description: Operador creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operador" }
        "400": { $ref: "#/components/responses/ErrorValidacion" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "403": { $ref: "#/components/responses/SinPermisos" }
        "409": { $ref: "#/components/responses/Conflicto" }
        "503": { $ref: "#/components/responses/BaseNoDisponible" }
  /api/operadores/{id_operador}:
    parameters:
      - name: id_operador
        in: path
        required: true
        description: Identificador que viaja en el JSON del nodo fisico.
        schema: { type: string, example: OP1 }
    get:
      tags: [Operadores]
      summary: Obtener un operador
      responses:
        "200":
          description: Operador
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operador" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
    put:
      tags: [Operadores]
      summary: Actualizar un operador
      description: |
        Solo administrador. Campos aceptados: `nombre`, `edad`, `genero`,
        `rol`, `metodo_acceso`, `activo`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OperadorActualizacion" }
      responses:
        "200":
          description: Operador actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operador" }
        "400": { $ref: "#/components/responses/ErrorValidacion" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "403": { $ref: "#/components/responses/SinPermisos" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
    delete:
      tags: [Operadores]
      summary: Eliminar un operador
      description: Solo administrador.
      responses:
        "200":
          description: Operador eliminado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "403": { $ref: "#/components/responses/SinPermisos" }
        "404": { $ref: "#/components/responses/NoEncontrado" }

  # --------------------------------------------------------------- Puerta ---
...
  /api/puerta/resumen:
    get:
      tags: [Puerta]
      summary: Conteo de veces abierta y cerrada
      description: |
        Publico. Alimenta los paneles "cantidad de veces abierta" y "cantidad de
        veces cerrada" de Grafana. Rango por defecto: ultimos 7 dias.
      security: []
      parameters:
        - $ref: "#/components/parameters/Desde"
        - $ref: "#/components/parameters/Hasta"
        - $ref: "#/components/parameters/IdOperadorQuery"
      responses:
        "200":
          description: Resumen del rango
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResumenPuerta" }

  # ------------------------------------------------------------- Reportes ---
...
  /api/historico/series:
    get:
      tags: [Historico]
      summary: Serie temporal agrupada por intervalos
      description: >
        Promedios de temperatura y humedad agrupados en bloques de tiempo, con
        el minimo y el maximo de cada bloque. La agregacion la hace MongoDB, no
        el cliente. Si el rango pedido genera mas de 5000 puntos, el intervalo
        se ensancha solo y el valor aplicado vuelve en la respuesta.
        Publico: lo consume Grafana.
      security: []
      parameters:
        - { name: desde, in: query, schema: { type: string, format: date-time }, description: "Inicio del rango. Por defecto, 24 h atras." }
        - { name: hasta, in: query, schema: { type: string, format: date-time }, description: "Fin del rango. Por defecto, ahora." }
        - { name: intervalo, in: query, schema: { type: integer, default: 5, minimum: 1, maximum: 1440 }, description: "Minutos por bloque." }
      responses:
        "200":
          description: Serie agregada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SerieTemporal" }
  /api/historico/diario:
    get:
      tags: [Historico]
      summary: Resumen por dia
      description: >
        Una fila por dia con promedios, extremos, cuantas lecturas cayeron en
        alarma y cuantas veces se abrio la puerta. Publico.
      security: []
      parameters:
        - { name: dias, in: query, schema: { type: integer, default: 7, minimum: 1, maximum: 365 } }
      responses:
        "200":
          description: Resumen diario
          content:
            application/json:
              schema:
                type: object
                properties:
                  dias: { type: integer }
                  filas: { type: integer }
                  resumen:
                    type: array
                    items: { $ref: "#/components/schemas/ResumenDiario" }
  /api/historico/alarmas:
    get:
      tags: [Historico]
      summary: Conteo de alarmas por tipo
      description: Cuantas alarmas hubo de cada tipo y cuanto duraron. Publico.
      security: []
      parameters:
        - { name: desde, in: query, schema: { type: string, format: date-time } }
        - { name: hasta, in: query, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Alarmas agrupadas por tipo
          content:
            application/json:
              schema:
                type: object
                properties:
                  desde: { type: string, format: date-time }
                  hasta: { type: string, format: date-time }
                  total: { type: integer }
                  por_tipo:
                    type: array
                    items: { $ref: "#/components/schemas/AlarmaPorTipo" }
  /api/historico/operadores:
    get:
      tags: [Historico]
      summary: Actividad de puerta por operador
      description: >
        Ranking de operadores por cantidad de aperturas, con el saldo de huevos
        de cada uno. Son totales agregados, no datos personales, asi que es
        publico; los informes individuales siguen protegidos en /api/perfil.
      security: []
      parameters:
        - { name: desde, in: query, schema: { type: string, format: date-time } }
        - { name: hasta, in: query, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: Actividad por operador
          content:
            application/json:
              schema:
                type: object
                properties:
                  desde: { type: string, format: date-time }
                  hasta: { type: string, format: date-time }
                  operadores:
                    type: array
                    items: { $ref: "#/components/schemas/ActividadOperador" }
  /api/historico/lote/{lote_id}:
    get:
      tags: [Historico]
      summary: Condiciones durante la incubacion de un lote
      description: >
        Las lecturas no llevan lote_id (el nodo mide la incubadora, no sabe que
        lote hay dentro), asi que el cruce se hace por tiempo: la ventana entre
        la fecha de ingreso del lote y su cierre, o ahora si sigue incubando.
        Publico.
      security: []
      parameters:
        - { name: lote_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Estadisticas del lote
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EstadisticasLote" }
        "404":
          description: Lote no encontrado o sin datos
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token de POST /api/auth/login. Cabecera: Authorization: Bearer <token>"
  parameters:
    Desde:
      name: desde
      in: query
      description: Inicio del rango, ISO-8601 (ej. 2026-09-01T00:00:00Z).
      schema: { type: string, format: date-time }
    Hasta:
      name: hasta
      in: query
      description: Fin del rango, ISO-8601. Por defecto, ahora.
      schema: { type: string, format: date-time }
...
    IdOperadorQuery:
      name: id_operador
      in: query
      description: |
        Filtrar por operador. Un operador autenticado no puede consultar otro
        distinto al suyo: el valor se ignora y se usa el propio.
      schema: { type: string, example: OP1 }
  responses:
    ErrorValidacion:
      description: El cuerpo de la peticion no paso las validaciones
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErroresValidacion" }
    NoAutenticado:
      description: Falta el token o no es valido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
    SinPermisos:
      description: El rol del token no alcanza para esta operacion
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
    NoEncontrado:
      description: El recurso no existe
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
    Conflicto:
      description: El recurso ya existe o la operacion no esta permitida
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
    BaseNoDisponible:
      description: |
        La base de datos no responde. El backend sigue en linea en modo
        degradado, pero no puede persistir.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
  schemas:
    Mensaje:
      type: object
      properties:
        mensaje: { type: string }
    ErroresValidacion:
      type: object
      properties:
        errores:
          type: array
          items: { type: string }
          example: ["El campo 'codigo' es obligatorio."]
    Rol:
      type: string
      enum: [administrador, operador_1, operador_2]
    Usuario:
      type: object
      properties:
        username: { type: string, example: operador1 }
        nombre: { type: string, example: Operador 1 }
        rol: { $ref: "#/components/schemas/Rol" }
        id_operador:
          type: string
          nullable: true
          description: Enlace con el operador del nodo fisico. Null para el administrador.
          example: OP1
        activo: { type: boolean }
    RespuestaLogin:
      type: object
      properties:
        token: { type: string, example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... }
        token_tipo: { type: string, example: Bearer }
        expira_en: { type: string, format: date-time }
        usuario: { $ref: "#/components/schemas/Usuario" }
    Perfil:
      type: object
      properties:
        usuario: { $ref: "#/components/schemas/Usuario" }
        es_administrador: { type: boolean }
        operador:
          allOf: [{ $ref: "#/components/schemas/Operador" }]
          nullable: true
        permisos:
          type: object
          properties:
            ver_todos_los_operadores: { type: boolean }
            administrar_usuarios: { type: boolean }
            administrar_operadores: { type: boolean }
    Lectura:
      type: object
      properties:
        id: { type: string }
        temperatura: { type: number, format: float, example: 37.5 }
        humedad: { type: number, format: float, example: 55.2 }
        sistema_encendido: { type: boolean }
        alarma: { type: boolean }
        espacios_libres: { type: integer, minimum: 0, maximum: 6 }
        calefaccion: { type: boolean }
        ventilacion: { type: boolean }
        rotacion: { type: boolean }
        puerta_abierta: { type: boolean }
        node_id: { type: string, example: incubadora-01 }
        timestamp: { type: string, format: date-time }
...
    LoteEntrada:
      type: object
      required: [codigo, nombre_lote, cantidad_huevos]
      properties:
        codigo: { type: string, example: L-001 }
        nombre_lote: { type: string, example: Lote criollo septiembre }
        cantidad_huevos: { type: integer, minimum: 1, example: 6 }
        raza: { type: string, example: Criolla }
        responsable: { type: string, example: Juan Ordonez }
        observaciones: { type: string }
        estado:
          type: string
          enum: [incubando, eclosionado, descartado]
          default: incubando
    LoteActualizacion:
      type: object
      description: Todos los campos son opcionales; `codigo` no se puede cambiar.
      properties:
        nombre_lote: { type: string }
        cantidad_huevos: { type: integer, minimum: 1 }
        raza: { type: string }
        responsable: { type: string }
        observaciones: { type: string }
        estado:
          type: string
          enum: [incubando, eclosionado, descartado]
    Lote:
      allOf:
        - type: object
          properties:
            id: { type: string }
            fecha_ingreso: { type: string, format: date-time }
        - $ref: "#/components/schemas/LoteEntrada"
    OperadorEntrada:
      type: object
      required: [id_operador, nombre]
      properties:
        id_operador:
          type: string
          description: Debe coincidir con el id que emite el firmware.
          example: OP1
        nombre: { type: string, example: Ana Lopez }
        edad: { type: integer, minimum: 0, maximum: 120, example: 34 }
        genero:
          type: string
          enum: [masculino, femenino, otro, no_especificado]
          default: no_especificado
        rol:
          type: string
          enum: [operador_1, operador_2, administrador]
          default: operador_1
        metodo_acceso:
          type: string
          enum: [rfid, huella, llave, manual]
          default: rfid
        activo: { type: boolean, default: true }
    OperadorActualizacion:
      type: object
      description: Todos opcionales; `id_operador` no se puede cambiar.
      properties:
        nombre: { type: string }
        edad: { type: integer, minimum: 0, maximum: 120 }
        genero:
          type: string
          enum: [masculino, femenino, otro, no_especificado]
        rol:
          type: string
          enum: [operador_1, operador_2, administrador]
        metodo_acceso:
          type: string
          enum: [rfid, huella, llave, manual]
        activo: { type: boolean }
    Operador:
      allOf:
        - type: object
          properties:
            id: { type: string }
            fecha_registro: { type: string, format: date-time }
        - $ref: "#/components/schemas/OperadorEntrada"
...
    ResumenPuerta:
      type: object
      properties:
        desde: { type: string, format: date-time }
        hasta: { type: string, format: date-time }
        veces_abierta: { type: integer }
        veces_cerrada: { type: integer }
        huevos_ingresados: { type: integer }
        huevos_retirados: { type: integer }
        por_operador:
          type: array
          items:
            type: object
            properties:
              id_operador: { type: string }
              veces_abierta: { type: integer }
              veces_cerrada: { type: integer }
              huevos_ingresados: { type: integer }
              huevos_retirados: { type: integer }
...
    SerieTemporal:
      type: object
      properties:
        desde: { type: string, format: date-time }
        hasta: { type: string, format: date-time }
        intervalo_minutos: { type: integer, description: "Intervalo realmente aplicado." }
        puntos: { type: integer }
        series:
          type: array
          items:
            type: object
            properties:
              timestamp: { type: string, format: date-time }
              temp_prom: { type: number, nullable: true }
              temp_min: { type: number, nullable: true }
              temp_max: { type: number, nullable: true }
              hum_prom: { type: number, nullable: true }
              hum_min: { type: number, nullable: true }
              hum_max: { type: number, nullable: true }
              lecturas: { type: integer }
              alarmas: { type: integer, description: "Lecturas del bloque con la bandera de alarma." }
    ResumenDiario:
      type: object
      properties:
        dia: { type: string, example: "2026-09-05" }
        temp_prom: { type: number, nullable: true }
        temp_min: { type: number, nullable: true }
        temp_max: { type: number, nullable: true }
        hum_prom: { type: number, nullable: true }
        hum_min: { type: number, nullable: true }
        hum_max: { type: number, nullable: true }
        lecturas: { type: integer }
        lecturas_en_alarma: { type: integer }
        aperturas_puerta: { type: integer }
    AlarmaPorTipo:
      type: object
      properties:
        tipo: { type: string }
        cantidad: { type: integer }
        duracion_total_segundos: { type: number }
        duracion_promedio_segundos: { type: number, nullable: true }
        activas: { type: integer }
    ActividadOperador:
      type: object
      properties:
        id_operador: { type: string }
        veces_abierta: { type: integer }
        veces_cerrada: { type: integer }
        huevos_ingresados: { type: integer }
        huevos_retirados: { type: integer }
    EstadisticasLote:
      type: object
      properties:
        lote_id: { type: string }
        codigo: { type: string, nullable: true }
        estado: { type: string, nullable: true }
        desde: { type: string, format: date-time }
        hasta: { type: string, format: date-time }
        dias_incubando: { type: number }
        temp_min: { type: number, nullable: true }
        temp_max: { type: number, nullable: true }
        temp_prom: { type: number, nullable: true }
        hum_min: { type: number, nullable: true }
        hum_max: { type: number, nullable: true }
        hum_prom: { type: number, nullable: true }
        lecturas_totales: { type: integer }